build: make the merge compile, and fix what the compiler exposed
First compile of this merge -- six files had been resolved without one. Most
errors were mechanical, three were merge content the compiler happened to catch.

Mechanical: base_to_precision's declaration and definition were resolved from
opposite sides (size_t prototype, uint8_t definition -- conflicting types under
C11 6.7.6.3p15); fsm.c lost the signed_metadata.h include; fsm_msg_solana.h
carried a 5-arg ed25519_sign against the pinned trezor-crypto's 4-arg form;
unittests/board/board.cpp had been left structurally unparseable by the merge
(duplicated include blocks, a dangling return, one test's body swallowing
another's). No assertion was touched repairing it.

Merge content the compiler exposed:

  fsm_msg_thorchain.h had develop's send block, which hardcodes " RUNE" and
  passes no denom, against alpha's 3-arg thorchain_signTxUpdateMsgSend. Restored
  alpha's block rather than hardcoding "rune" to satisfy the arity -- that would
  have signed a host-chosen denom while displaying a fixed one.

  authenticator.c had been taken wholesale from develop, which is strictly older
  than alpha's here and drops, in one go: authDisplayFieldValid() (non-printable
  and empty domain/account refused before display), AUTHSECRET_SIZE_MIN
  (brute-forceable TOTP secrets), duplicate-identity rejection, the split "TOTP
  Secret" screen that closes the OLED body overflow, removal of ALL duplicate
  slots on delete, and memzero(authSecret) on every exit path. develop's only
  contribution over the base was CANCELED plus unsigned wipeAuthData(), which
  alpha already has as AUTH_CANCELLED. Restored alpha's. Renaming the enum in
  the test would have compiled while leaving those five checks deleted.

Two toolchain traps, neither a source problem:

  nanopb MUST be 0.3.9.4. 0.4.x parses the plugin parameter as
  `if ',' not in params and ' -' in params`; the repo spells it
  `--nanopb_out=-f types.options:.` which has neither, so the whole string lexes
  as one token, every .options file is silently ignored, and the build dies at
  the repo's own `! grep -r pb_callback_t` guard with no hint why.

  -DPB_NO_PACKED_STRUCTS=1 is mandatory on arm64, else the link fails with
  "pointer not aligned in '_HDNodePathType_fields'".

Also: the local note that firmware-unit hangs on this Mac is STALE for this
tree. All six suites it names run to completion in milliseconds.

Not fixed here, deliberately: recovery_cipher.c lost alpha's "previous word"
indicator (passing NULL to compile). develop's file is not simply older -- it
carries the #429 staging rework -- so replaying alpha's prev_info block onto it
is a real merge, not a revert.

+++ lib/board/util.c
@@ -104,7 +104,7 @@ bool is_valid_ascii(const uint8_t* data, uint32_t size) {
 
 /* convert number in base units to specified decimal precision */
 int base_to_precision(uint8_t* dest, const uint8_t* value,
-                      const uint8_t dest_len, const uint8_t value_len,
+                      const size_t dest_len, const size_t value_len,
                       const uint8_t precision) {
   if (!(dest && value)) {
     // invalid pointer
